Introducing CANopyner
CANopyner is an effort to produce a Python based tool for interaction
with CANopen devices and networks.  The initial focus is on access to a
device's object dictionary.  Presently an EDS file can be parsed and
it's indexes and subindexes can be displayed in tree.

The .idea directory contains the PyCharm project files.

Thanks to the following source for helping me get started.  See the
LICENSE file for more info.

*
http://www.riverbankcomputing.com/pipermail/pyqt/2009-April/022729.html